🚚 [Rename] 폴더명 변경 todo -> todo_detail
