+++ lib/ui/todo/screen/todo_add_screen.dart
@@ -4,9 +4,9 @@ import 'package:sqflite/sqflite.dart';
 import 'package:time_todo/bloc/category_detail/category_detail_bloc.dart';
 import 'package:time_todo/bloc/category_detail/category_detail_event.dart';
 import 'package:time_todo/bloc/category_detail/category_detail_state.dart';
-import 'package:time_todo/bloc/todo/todo_detail_bloc.dart';
-import 'package:time_todo/bloc/todo/todo_detail_event.dart';
-import 'package:time_todo/bloc/todo/todo_detail_state.dart';
+import 'package:time_todo/bloc/todo_detail/todo_detail_bloc.dart';
+import 'package:time_todo/bloc/todo_detail/todo_detail_event.dart';
+import 'package:time_todo/bloc/todo_detail/todo_detail_state.dart';
 import 'package:time_todo/ui/components/widget/date_picker.dart';
 import 'package:time_todo/ui/components/widget/time_picker.dart';
 import 'package:time_todo/ui/components/widget/toast_message.dart';
