+++ lib/ui/todo/screen/todo_modify_screen.dart
@@ -3,14 +3,14 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:time_todo/bloc/category_detail/category_detail_bloc.dart';
 import 'package:time_todo/bloc/category_detail/category_detail_event.dart';
 import 'package:time_todo/bloc/category_detail/category_detail_state.dart';
-import 'package:time_todo/bloc/todo/todo_detail_bloc.dart';
-import 'package:time_todo/bloc/todo/todo_detail_state.dart';
+import 'package:time_todo/bloc/todo_detail/todo_detail_bloc.dart';
+import 'package:time_todo/bloc/todo_detail/todo_detail_event.dart';
+import 'package:time_todo/bloc/todo_detail/todo_detail_state.dart';
 import 'package:time_todo/ui/components/buttons/main_delete_button.dart';
 import 'package:time_todo/ui/components/widget/main_alert.dart';
 import 'package:time_todo/ui/components/widget/time_picker.dart';
 import 'package:time_todo/ui/utils/date_time_utils.dart';
 import 'package:time_todo/ui/utils/debouncer.dart';
-import '../../../bloc/todo/todo_detail_event.dart';
 import '../../../entity/todo/todo_tbl.dart';
 import '../../components/widget/date_picker.dart';
 import '../../components/widget/main_app_bar.dart';
